[GNU-EFI]:Compilation support for old compilers.
BMDA-1537

Compilation should not break with old compilers and -DGNU_EFI_USE_MS_ABI.
It's should be allowed to request GNU_EFI_USE_MS_ABI even if the current
compiler doesn't support it. With this type checking will not work but atleast
compilation should not break and gnu-efi falls back to
using legacy ways of setting the calling convention.

Change-Id: I1675336eae26e9098fac9adba5f811fed49cc946
Signed-off-by: sahashau <shaunak.saha@intel.com>
Reviewed-on: https://android.intel.com/289627
Reviewed-by: sys_pccgand <sys_pccgand@intel.com>
Reviewed-by: Sripathi, Srinivas <srinivas.sripathi@intel.com>
Reviewed-by: Kandru, Suneel <suneel.kandru@intel.com>